Move some common UI to BaseWidgetConfigActivity.kt
